backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
numpy==2.1.1
faiss-cpu==1.8.0.post1
requests==2.32.3
